Refactor File Uploads (#63)
refs: MBL-12078
affects: student
release note: none

+++ Core/CoreTests/Extensions/Int64ExtensionsTests.swift
@@ -19,6 +19,6 @@ import XCTest
 
 class Int64ExtensionsTests: XCTestCase {
     func testHumanReadableFileSize() {
-        XCTAssertEqual(Int64(1).humanReadableFileSize, "1 byte")
+        XCTAssertEqual(1.humanReadableFileSize, "1 byte")
     }
 }
